refactor(crm): rename engine/crm/constants.py to expressions.py
After the Step 1-3 refactor, engine/crm/constants.py no longer holds
any regulatory constants or input-domain type sets — those moved to
data/tables/ and data/schemas.py respectively. The file now contains
only Polars expression builders (collateral_lgd_expr,
overcollateralisation_ratio_expr, min_collateralisation_threshold_expr,
is_financial_collateral_type_expr, collateral_category_expr,
beneficiary_level_expr), a supervisory_lgd_values helper, and two
small data tables driving the Art. 231 waterfall (WATERFALL_ORDER,
CRM_ALLOC_COLUMNS).

The name "constants" is misleading. Rename to expressions.py to match
the dominant content; the small waterfall data tables stay
co-located because they're the config the expression builders
implement (classification expressions and waterfall ordering are one
concern).

Mechanics:
- git mv preserves file history.
- Module docstring rewritten to describe expression builders + Art.
  231 waterfall mechanics, with cross-refs to the sibling data
  modules it consumes.
- Five import sites updated (collateral.py, processor.py deferred
  import, three test files).
- Stale comment in data/schemas.py:514 that referenced the old
  filename also updated.
