+++ src/rwa_calc/engine/crm/collateral.py
@@ -30,7 +30,7 @@ import polars as pl
 from rwa_calc.data.schemas import NON_ELIGIBLE_RE_TYPES
 from rwa_calc.data.tables.crm_supervisory import MIN_COLLATERALISATION_THRESHOLDS
 from rwa_calc.domain.enums import AIRBCollateralMethod, ApproachType
-from rwa_calc.engine.crm.constants import (
+from rwa_calc.engine.crm.expressions import (
     CRM_ALLOC_COLUMNS,
     WATERFALL_ORDER,
     beneficiary_level_expr,
